+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter } from 'react-router-dom';
-import { AppRouter } from './components/AppRouter';
+import { AppRouter } from './components';
 
 function App() {
   return (
@@ -9,4 +9,4 @@ function App() {
   );
 }
 
-export { App };
+export default App;
