+++ src/components/LoginForm.tsx
@@ -4,4 +4,4 @@ const LoginForm = () => {
   return <></>;
 };
 
-export { LoginForm };
+export default LoginForm;
