+++ src/components/Logo.tsx
@@ -12,4 +12,4 @@ const Logo = () => {
   );
 };
 
-export { Logo };
+export default Logo;
