+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as AppRouter } from './AppRouter';
+export { default as GraphiQL } from './GraphiQL';
+export { default as LoginForm } from './LoginForm';
+export { default as Logo } from './Logo';
+export { default as RegistrationForm } from './RegistrationForm';
