+++ src/layout/Footer.tsx
@@ -6,4 +6,4 @@ const Footer = () => {
   );
 };
 
-export { Footer };
+export default Footer;
