+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { App } from './App';
+import App from './App';
 import './assets/scss/style.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
