+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as Login } from './Login';
+export { default as Main } from './Main';
+export { default as NotFound } from './NotFound';
+export { default as Registration } from './Registration';
